+++ modules/juce_graphics/fonts/juce_GlyphArrangement.h
@@ -307,9 +307,9 @@ private:
     int fitLineIntoSpace (int start, int numGlyphs, float x, float y, float w, float h, const Font&,
                           Justification, float minimumHorizontalScale);
     void spreadOutLine (int start, int numGlyphs, float targetWidth);
-    void splitLines (const String&, Font, int start, int x, int y, int w, int h, int maxLines,
+    void splitLines (const String&, Font, int start, float x, float y, float w, float h, int maxLines,
                      float lineWidth, Justification, float minimumHorizontalScale);
-    void addLinesWithLineBreaks (const String&, const Font&, int x, int y, int width, int height, Justification);
+    void addLinesWithLineBreaks (const String&, const Font&, float x, float y, float width, float height, Justification);
     void drawGlyphUnderline (const Graphics&, const PositionedGlyph&, int, const AffineTransform&) const;
 
     JUCE_LEAK_DETECTOR (GlyphArrangement)
